publishing of the convention plugins
